requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.0.5
cryptography==41.0.7
twilio==8.10.0
python-dotenv==1.0.0
Pillow==10.0.1
numpy==1.24.3
opencv-python==4.8.1.78
tensorflow==2.13.0
mediapipe==0.10.9 